Regular expression corrected

+++ src/main/java/com/amisoft/utils/ConversionUtility.java
@@ -3,6 +3,7 @@ package com.amisoft.utils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
@@ -12,21 +13,21 @@ import java.util.Map;
 public class ConversionUtility {
 
 
-    public void loadmainCurrencyConversion(String mainConversionTable,Map<String,Double> targeteMap){
+    public void loadmainCurrencyConversion(String mainConversionTable,Map<String,BigDecimal> targeteMap){
 
         LoadMapFromProperty(mainConversionTable, targeteMap);
     }
 
 
 
-    public void LoadMapFromProperty(String mainConversionTable, Map<String, Double> targeteMap) {
+    public void LoadMapFromProperty(String mainConversionTable, Map<String, BigDecimal> targeteMap) {
 
         List<String> mainConversionRateList = Arrays.asList(StringUtils.split(mainConversionTable,","));
 
         mainConversionRateList.forEach(convRatePair -> {
 
             String[] currencyConv = StringUtils.split(convRatePair,"=");
-            targeteMap.putIfAbsent(currencyConv[0], Double.valueOf(currencyConv[1]));
+            targeteMap.putIfAbsent(currencyConv[0], BigDecimal.valueOf(Double.valueOf(currencyConv[1])));
         });
 
         System.out.println(targeteMap.keySet().toString());
